Included comments, timing of Kernels, ...

+++ 3D/CUDA/CUDA_NBody/include/Constants.h
@@ -10,7 +10,7 @@ typedef struct SimulationParameters
     bool debug;
     bool benchmark;
     bool fullscreen;
-    float iterations;
+    int iterations;
     float timestep;
     float gravity;
     float dampening;
